test-streamer: make class properties public
